[refactor] :recycle: relocate publicRoute function to its sole usage location (#276)
