+++ src/api/pingRoutes.ts
@@ -1,4 +1,3 @@
-
 import { FastifyInstance } from 'fastify';
 
 /**
